Add simple ko test to the game tests #54
